Fixed to make sure all data can be read from T1932
It turns out that the time between consecutive read and writes from and to the T1932 need to controlled tightly to make sure that all 48 bytes are read (incl. telemetry updates for speed, cadence, etc.). In case not all bytes are read, keep these values unchanged iso changing to zero.